Fix PR to synchronize with recent work
This is mostly just me resubmitting the PR with the various naming and synchronizations already known. I also fixed up a few small logic issues in Pinnacle Rock and the moon. Still probably undertested, but it runs and makes seeds that look like they are using what it's supposed to do correctly.