Invisible Soldier Logic and Can Get Potions Macros
Shiro was missing Soul Logic, as well as potion access was never updated to account for NPC souls being shuffled since it still assumed access to the free red potion from Kotake. Tested with a handful of generations with the check set as a priority location. All of these attempts showed the expected/added requirements in spheres prior to the check. Also added a macro for red/blue potion access to not have ~30 lines of logic for one check